+++ src/app/events/pages/events-details/events-details.ts
@@ -22,7 +22,7 @@ import { Component } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
 
-import { EventItem } from 'src/app/entities/eventItem';
+import { EventItem } from 'src/app/models/eventItem';
 import { AlertService } from 'src/app/services/utils-services/alert-service';
 import { UserService } from 'src/app/services/utils-services/user-service';
