+++ src/app/libraries/pages/library-details/library-details.ts
@@ -22,7 +22,7 @@ import { animate, state, style, transition, trigger } from '@angular/animations'
 import { Component } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 
-import { LibraryItem } from 'src/app/entities/libraryItem';
+import { LibraryItem } from 'src/app/models/libraryItem';
 import { UtilsService } from 'src/app/services/utils-services/utils-services';
 import { LibrariesService } from 'src/app/services/wso2-services/libraries-service';
