+++ src/app/models/eventItem.ts
