+++ src/app/models/libraryItem.ts
