+++ src/app/models/newsItem.ts
