+++ src/app/models/timeSlot.ts
