+++ src/app/services/map-services/map-service.ts
@@ -24,7 +24,7 @@ import { Injectable } from '@angular/core';
 import { Geolocation } from '@ionic-native/geolocation/ngx';
 import { MenuController, Platform } from '@ionic/angular';
 
-import { MapLocation } from 'src/app/entities/mapLocation';
+import { MapLocation } from 'src/app/models/mapLocation';
 import { ConnectivityService } from '../utils-services/connectivity-service';
 import { UserService } from '../utils-services/user-service';
