+++ src/app/services/rss-services/events-service.ts
@@ -20,7 +20,7 @@
  */
 import { Injectable } from '@angular/core';
 
-import { EventItem } from 'src/app/entities/eventItem';
+import { EventItem } from 'src/app/models/eventItem';
 import { UserService } from '../utils-services/user-service';
 import { RssService } from './rss-service';
