+++ src/app/services/wso2-services/repertoire-service.ts
@@ -21,7 +21,7 @@ import { HttpClient } from '@angular/common/http';
  */
 import { Injectable } from '@angular/core';
 
-import { EmployeeItem } from 'src/app/entities/employeeItem';
+import { EmployeeItem } from 'src/app/models/employeeItem';
 import { ConnectivityService } from '../utils-services/connectivity-service';
 import { Wso2Service } from './wso2-service';
