+++ src/app/studies/pages/course/course.ts
@@ -26,8 +26,8 @@ import { Calendar } from '@ionic-native/calendar/ngx';
 import { AlertController, IonItemSliding, ModalController } from '@ionic/angular';
 import { TranslateService } from '@ngx-translate/core';
 
-import { Activity } from 'src/app/entities/activity';
-import { Course } from 'src/app/entities/course';
+import { Activity } from 'src/app/models/activity';
+import { Course } from 'src/app/models/course';
 import { CourseService } from 'src/app/services/studies-services/course-service';
 import { UserService } from 'src/app/services/utils-services/user-service';
 import { ModalInfoPage } from './modal-info/modal-info';
