+++ src/app/support/pages/support.ts
@@ -24,7 +24,7 @@ import { animate, state, style, transition, trigger } from '@angular/animations'
 import { Component } from '@angular/core';
 import { InAppBrowser } from '@ionic-native/in-app-browser/ngx';
 
-import { EmployeeItem } from 'src/app/entities/employeeItem';
+import { EmployeeItem } from 'src/app/models/employeeItem';
 import { RepertoireService } from 'src/app/services/wso2-services/repertoire-service';
 
 @Component({
